security/manager/ssl/tests/mochitest/bugs/test_bug234856.html

<!doctype html>
<html><head>
  <meta charset="utf-8">
  <title>Test for Bug 234856</title>
  <script src="chrome://mochikit/content/MochiKit/packed.js"></script>
  <script src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"
  ></script>
  <link rel="stylesheet"
        href="chrome://mochikit/content/tests/SimpleTest/test.css">
</head><body>
  <script>
    const Ci = Components.interfaces;
    SimpleTest.waitForExplicitFinish();
    window.onload = function RunTest()
    {
      var certDB = Components.classes["@mozilla.org/security/x509certdb;1"]
                   .getService(Ci.nsIX509CertDB);

      var rawCert = document.getElementById("cert").firstChild.wholeText;
      var cert = certDB.constructX509FromBase64(rawCert);

      // all simple attributes that are properties of the cert itself
      is(cert.certType, Ci.nsIX509Cert.EMAIL_CERT, "certType");
      is(cert.isSelfSigned, false, "isSelfSigned");
      is(cert.serialNumber, "00:98:99:6F", "serialNumber");
      is(cert.sha1Fingerprint, "03:E2:EB:66:DC:D3:44:10:F5:FE:9E:C7:8D:51:3C:BF:9A:AD:7C:5E", "sha1Fingerprint");
      is(cert.sha256Fingerprint, "A5:CE:5B:9C:F1:56:3D:D6:8D:03:C5:A0:6C:9E:D3:10:01:B3:81:C9:CF:21:5F:E4:D2:D4:A9:73:C4:8C:9E:30", "sha256Fingerprint");

      is(cert.nickname, "(no nickname)", "nickname");
      is(cert.emailAddress, "ludek.rasek@centrum.cz", "emailAddress");
      is(cert.subjectName, "serialNumber=ICA - 10003769,SN=Rašek,name=Luděk Rašek,initials=LR,givenName=Luděk,E=ludek.rasek@centrum.cz,L=\"Pacov, Nádražní 769\",ST=Vysočina,CN=Luděk Rašek,C=CZ", "subjectName");
      is(cert.commonName, "Luděk Rašek", "commonName");
      is(cert.organization, "", "organization");
      is(cert.organizationalUnit, "", "organizationalUnit");
      is(cert.windowTitle, "Luděk Rašek", "windowTitle");

      is(cert.issuerName, "OU=Akreditovaný poskytovatel certifikačních služeb,O=První certifikační autorita a.s.,L=\"Podvinný mlýn 2178/6, 190 00 Praha 9\",C=CZ,CN=I.CA - Qualified root certificate (kvalifikovaný certifikát poskytovatele) - PSEUDONYM", "issuerName");
      is(cert.issuerCommonName, "I.CA - Qualified root certificate (kvalifikovaný certifikát poskytovatele) - PSEUDONYM", "issuerCommonName");
      is(cert.issuerOrganization, "První certifikační autorita a.s.", "issuerOrganization");
      is(cert.issuerOrganizationUnit, "Akreditovaný poskytovatel certifikačních služeb", "issuerOrganizationUnit");

      SimpleTest.finish();
    }
  </script>
  <div style="display:none" id="cert">MIIF3DCCBMSgAwIBAgIEAJiZbzANBgkqhkiG9w0BAQUFADCCAQ0xYTBfBgNVBAMMWEkuQ0EgLSBRdWFsaWZpZWQgcm9vdCBjZXJ0aWZpY2F0ZSAoa3ZhbGlmaWtvdmFuw70gY2VydGlmaWvDoXQgcG9za3l0b3ZhdGVsZSkgLSBQU0VVRE9OWU0xCzAJBgNVBAYTAkNaMS8wLQYDVQQHDCZQb2R2aW5uw70gbWzDvW4gMjE3OC82LCAxOTAgMDAgUHJhaGEgOTEsMCoGA1UECgwjUHJ2bsOtIGNlcnRpZmlrYcSNbsOtIGF1dG9yaXRhIGEucy4xPDA6BgNVBAsMM0FrcmVkaXRvdmFuw70gcG9za3l0b3ZhdGVsIGNlcnRpZmlrYcSNbsOtY2ggc2x1xb5lYjAeFw0wMjEyMTIxMzMzNDZaFw0wMzEyMTIxMzMzNDZaMIIBFDELMAkGA1UEBhMCQ1oxHzAdBgNVBAMeFgBMAHUAZAEbAGsAIABSAGEBYQBlAGsxGTAXBgNVBAgeEABWAHkAcwBvAQ0AaQBuAGExLzAtBgNVBAceJgBQAGEAYwBvAHYALAAgAE4A4QBkAHIAYQF+AG4A7QAgADcANgA5MSUwIwYJKoZIhvcNAQkBFhZsdWRlay5yYXNla0BjZW50cnVtLmN6MRMwEQYDVQQqHgoATAB1AGQBGwBrMQ0wCwYDVQQrHgQATABSMR8wHQYDVQQpHhYATAB1AGQBGwBrACAAUgBhAWEAZQBrMRMwEQYDVQQEHgoAUgBhAWEAZQBrMRcwFQYDVQQFEw5JQ0EgLSAxMDAwMzc2OTCBnzANBgkqhkiG9w0BAQEFAAOBjQAwgYkCgYEAxc7dGd0cNlHZ7tUUl5k30bfYlY3lnOD049JGbTXSt4jNFMRLj6s/777W3kcIdcIwdKxjQULBKgryDvZJ1DAWp2TwzhPDVYj3sU4Niqb7mOUcp/4ckteUxGF6FmXtJR9+XHTuLZ+omF9HOUefheBKnXvZuqrLM16ynbJn4sPwwdcCAwEAAaOCAbswggG3MCUGA1UdEQQeMBygGgYKKwMGAQQB3BkCAaAMDAoxNzYyODk2ODgzMGkGA1UdHwRiMGAwHqAcoBqGGGh0dHA6Ly9xLmljYS5jei9xaWNhLmNybDAeoBygGoYYaHR0cDovL2IuaWNhLmN6L3FpY2EuY3JsMB6gHKAahhhodHRwOi8vci5pY2EuY3ovcWljYS5jcmwwHwYDVR0jBBgwFoAUK1oKfvvlDYUsZTByvGN701mca/UwHQYDVR0OBBYEFPAs70DB+LS0PnA6niPUfJ5wdQH5MIG4BgNVHSAEgbAwga0wgaoGCysGAQQBs2EBAQQEMIGaMC8GCCsGAQUFBwIBFiNodHRwOi8vd3d3LmljYS5jei9xY3AvY3BxcGljYTAyLnBkZjBnBggrBgEFBQcCAjBbGllUZW50byBjZXJ0aWZpa2F0IGplIHZ5ZGFuIGpha28gS3ZhbGlmaWtvdmFueSBjZXJ0aWZpa2F0IHYgc291bGFkdSBzZSB6YWtvbmVtIDIyNy8yMDAwIFNiLjAYBggrBgEFBQcBAwQMMAowCAYGBACORgEBMA4GA1UdDwEB/wQEAwIE8DANBgkqhkiG9w0BAQUFAAOCAQEAv2V+nnYYMIgabmmgHx49CtlZIHdGS3TuWKXw130xFhbXDnNhEbx3alaskNsvjQQRLqs1ZwKy58yynse+eJYHqenmHDACpAfVpCF9PXC/mDarVsoQw7NTcUpsAFhSd/zTv9jIf3twECyxx/RVzONVcob7nPePESHiKoG4FbtcuUh0wSHvCmTwRIQqPDCIuHcFStSt3Jr9iXcbXEhe4mSccOZ8N+r7Rv3ncKcevlRl7uFfDKDTyd43SZeRS/7J8KRfhD/h2nawrCFwc5gJW10aLJGFL/mcS7ViAIT9HCVk23j4TuBjsVmnZ0VKxB5edux+LIEqtU428UVHZWU/I5ngLw==</div>
</body></html>
